move `KwPair` and company to `core/`
